driver: npmx: shell: move ldsw code to the separate file
