+++ drivers/npmx/npmx_shell.c
@@ -14,12 +14,6 @@
 #include <math.h>
 #include <stdlib.h>
 
-/** @brief Load switch soft-start configuration parameter. */
-typedef enum {
-	LDSW_SOFT_START_PARAM_ENABLE, /* Soft-start enable. */
-	LDSW_SOFT_START_PARAM_CURRENT /* Soft-start current. */
-} ldsw_soft_start_config_param_t;
-
 /** @brief POF config parameter. */
 typedef enum {
 	POF_CONFIG_PARAM_POLARITY, /* POF polarity. */
@@ -47,14 +41,6 @@ typedef enum {
 } timer_config_param_t;
 
 static const struct device *pmic_dev = DEVICE_DT_GET(DT_NODELABEL(npm_0));
-
-static npmx_ldsw_t *ldsw_instance_get(const struct shell *shell, uint32_t index)
-{
-	npmx_instance_t *npmx_instance = npmx_instance_get(shell);
-	bool status = check_instance_index(shell, "LDSW", index, NPM_LDSW_COUNT);
-
-	return (npmx_instance && status) ? npmx_ldsw_get(npmx_instance, (uint8_t)index) : NULL;
-}
 
 static npmx_led_t *led_instance_get(const struct shell *shell, uint32_t index)
 {
@@ -90,485 +76,6 @@ static npmx_vbusin_t *vbusin_instance_get(const struct shell *shell)
 	npmx_instance_t *npmx_instance = npmx_instance_get(shell);
 
 	return npmx_instance ? npmx_vbusin_get(npmx_instance, 0) : NULL;
-}
-
-static int cmd_ldsw_active_discharge_set(const struct shell *shell, size_t argc, char **argv)
-{
-	args_info_t args_info = { .expected_args = 2,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-					  [1] = { SHELL_ARG_TYPE_BOOL_VALUE, "status" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	bool discharge_status = args_info.arg[1].result.bvalue;
-	npmx_error_t err_code =
-		npmx_ldsw_active_discharge_enable_set(ldsw_instance, discharge_status);
-
-	if (!check_error_code(shell, err_code)) {
-		print_set_error(shell, "LDSW active discharge status");
-		return 0;
-	}
-
-	print_success(shell, discharge_status, UNIT_TYPE_NONE);
-	return 0;
-}
-
-static int cmd_ldsw_active_discharge_get(const struct shell *shell, size_t argc, char **argv)
-{
-	args_info_t args_info = { .expected_args = 1,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	bool discharge_status;
-	npmx_error_t err_code =
-		npmx_ldsw_active_discharge_enable_get(ldsw_instance, &discharge_status);
-	if (!check_error_code(shell, err_code)) {
-		print_get_error(shell, "LDSW active discharge status");
-		return 0;
-	}
-
-	print_value(shell, discharge_status, UNIT_TYPE_NONE);
-	return 0;
-}
-
-static npmx_ldsw_gpio_t ldsw_gpio_index_convert(int32_t gpio_idx)
-{
-	switch (gpio_idx) {
-	case (-1):
-		return NPMX_LDSW_GPIO_NC;
-	case (0):
-		return NPMX_LDSW_GPIO_0;
-	case (1):
-		return NPMX_LDSW_GPIO_1;
-	case (2):
-		return NPMX_LDSW_GPIO_2;
-	case (3):
-		return NPMX_LDSW_GPIO_3;
-	case (4):
-		return NPMX_LDSW_GPIO_4;
-	default:
-		return NPMX_LDSW_GPIO_INVALID;
-	}
-}
-
-static int cmd_ldsw_gpio_set(const struct shell *shell, size_t argc, char **argv)
-{
-	args_info_t args_info = { .expected_args = 3,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-					  [1] = { SHELL_ARG_TYPE_INT32_VALUE, "GPIO number" },
-					  [2] = { SHELL_ARG_TYPE_BOOL_VALUE,
-						  "GPIO inversion status" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	int gpio_index = args_info.arg[1].result.ivalue;
-	npmx_ldsw_gpio_config_t gpio_config = {
-		.gpio = ldsw_gpio_index_convert(gpio_index),
-		.inverted = args_info.arg[2].result.bvalue,
-	};
-
-	if (gpio_config.gpio == NPMX_LDSW_GPIO_INVALID) {
-		shell_error(shell, "Error: wrong GPIO index.");
-		return 0;
-	}
-
-	if (!check_pin_configuration_correctness(shell, gpio_index)) {
-		return 0;
-	}
-
-	npmx_error_t err_code = npmx_ldsw_enable_gpio_set(ldsw_instance, &gpio_config);
-	if (!check_error_code(shell, err_code)) {
-		print_set_error(shell, "GPIO config");
-		return 0;
-	}
-
-	shell_print(shell, "Success: %d %d.", gpio_index, gpio_config.inverted);
-	return 0;
-}
-
-static int cmd_ldsw_gpio_get(const struct shell *shell, size_t argc, char **argv)
-{
-	args_info_t args_info = { .expected_args = 1,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	npmx_ldsw_gpio_config_t gpio_config;
-	npmx_error_t err_code = npmx_ldsw_enable_gpio_get(ldsw_instance, &gpio_config);
-	if (!check_error_code(shell, err_code)) {
-		print_get_error(shell, "GPIO config");
-		return 0;
-	}
-
-	int8_t gpio_index =
-		(gpio_config.gpio == NPMX_LDSW_GPIO_NC) ? -1 : ((int8_t)gpio_config.gpio - 1);
-
-	shell_print(shell, "Value: %d %d.", gpio_index, gpio_config.inverted);
-	return 0;
-}
-
-static int cmd_ldsw_ldo_voltage_set(const struct shell *shell, size_t argc, char **argv)
-{
-	args_info_t args_info = { .expected_args = 2,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-					  [1] = { SHELL_ARG_TYPE_UINT32_VALUE, "voltage" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	uint32_t voltage_mv = args_info.arg[1].result.uvalue;
-	npmx_ldsw_voltage_t ldsw_voltage = npmx_ldsw_voltage_convert(voltage_mv);
-	if (ldsw_voltage == NPMX_LDSW_VOLTAGE_INVALID) {
-		print_convert_error(shell, "millivolts", "LDSW voltage");
-		return 0;
-	}
-
-	npmx_error_t err_code = npmx_ldsw_ldo_voltage_set(ldsw_instance, ldsw_voltage);
-	if (!check_error_code(shell, err_code)) {
-		print_set_error(shell, "LDSW voltage");
-		return 0;
-	}
-
-	print_success(shell, voltage_mv, UNIT_TYPE_MILLIVOLT);
-	return 0;
-}
-
-static int cmd_ldsw_ldo_voltage_get(const struct shell *shell, size_t argc, char **argv)
-{
-	args_info_t args_info = { .expected_args = 1,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	npmx_ldsw_voltage_t ldsw_voltage;
-	npmx_error_t err_code = npmx_ldsw_ldo_voltage_get(ldsw_instance, &ldsw_voltage);
-	if (!check_error_code(shell, err_code)) {
-		print_get_error(shell, "LDSW voltage");
-		return 0;
-	}
-
-	uint32_t voltage_mv;
-	if (!npmx_ldsw_voltage_convert_to_mv(ldsw_voltage, &voltage_mv)) {
-		print_convert_error(shell, "LDSW voltage", "millivolts");
-		return 0;
-	}
-
-	print_value(shell, voltage_mv, UNIT_TYPE_MILLIVOLT);
-	return 0;
-}
-
-static int cmd_ldsw_mode_set(const struct shell *shell, size_t argc, char **argv)
-{
-	args_info_t args_info = {
-		.expected_args = 2,
-		.arg = { [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-			 [1] = { SHELL_ARG_TYPE_UINT32_VALUE, "mode" }, },
-	};
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	uint32_t mode = args_info.arg[1].result.uvalue;
-	npmx_ldsw_mode_t ldsw_mode;
-	switch (mode) {
-	case 0:
-		ldsw_mode = NPMX_LDSW_MODE_LOAD_SWITCH;
-		break;
-	case 1:
-		ldsw_mode = NPMX_LDSW_MODE_LDO;
-		break;
-	default:
-		shell_error(shell, "Error: Wrong mode:");
-		print_hint_error(shell, 0, "LOADSW");
-		print_hint_error(shell, 1, "LDO");
-		return 0;
-	}
-
-	npmx_error_t err_code = npmx_ldsw_mode_set(ldsw_instance, ldsw_mode);
-	if (err_code != NPMX_SUCCESS) {
-		print_set_error(shell, "LDSW mode");
-		return 0;
-	}
-
-	/* LDSW reset is required to apply mode change. */
-	err_code = npmx_ldsw_task_trigger(ldsw_instance, NPMX_LDSW_TASK_DISABLE);
-	if (!check_error_code(shell, err_code)) {
-		shell_error(shell, "Error: reset error while disabling LDSW to change mode.");
-		return 0;
-	}
-
-	err_code = npmx_ldsw_task_trigger(ldsw_instance, NPMX_LDSW_TASK_ENABLE);
-	if (!check_error_code(shell, err_code)) {
-		shell_error(shell, "Error: reset error while enabling LDSW to change mode.");
-		return 0;
-	}
-
-	print_success(shell, mode, UNIT_TYPE_NONE);
-	return 0;
-}
-
-static int cmd_ldsw_mode_get(const struct shell *shell, size_t argc, char **argv)
-{
-	args_info_t args_info = { .expected_args = 1,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	npmx_ldsw_mode_t mode;
-	npmx_error_t err_code = npmx_ldsw_mode_get(ldsw_instance, &mode);
-	if (!check_error_code(shell, err_code)) {
-		print_get_error(shell, "LDSW mode");
-		return 0;
-	}
-
-	print_value(shell, (int)mode, UNIT_TYPE_NONE);
-	return 0;
-}
-
-static int ldsw_soft_start_config_set(const struct shell *shell, size_t argc, char **argv,
-				      ldsw_soft_start_config_param_t config_type)
-{
-	shell_arg_type_t arg_type;
-	switch (config_type) {
-	case LDSW_SOFT_START_PARAM_ENABLE:
-		arg_type = SHELL_ARG_TYPE_BOOL_VALUE;
-		break;
-	case LDSW_SOFT_START_PARAM_CURRENT:
-		arg_type = SHELL_ARG_TYPE_UINT32_VALUE;
-		break;
-	}
-
-	args_info_t args_info = { .expected_args = 2,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-					  [1] = { arg_type, "config" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	npmx_ldsw_soft_start_config_t soft_start_config;
-	npmx_error_t err_code = npmx_ldsw_soft_start_config_get(ldsw_instance, &soft_start_config);
-	if (!check_error_code(shell, err_code)) {
-		print_get_error(shell, "soft-start config");
-		return 0;
-	}
-
-	shell_arg_result_t result = args_info.arg[1].result;
-	switch (config_type) {
-	case LDSW_SOFT_START_PARAM_ENABLE:
-		soft_start_config.enable = result.bvalue;
-		break;
-	case LDSW_SOFT_START_PARAM_CURRENT:
-		soft_start_config.current = npmx_ldsw_soft_start_current_convert(result.uvalue);
-		if (soft_start_config.current == NPMX_LDSW_SOFT_START_CURRENT_INVALID) {
-			print_convert_error(shell, "milliamperes", "soft-start current");
-			return 0;
-		}
-		break;
-	}
-
-	err_code = npmx_ldsw_soft_start_config_set(ldsw_instance, &soft_start_config);
-	if (!check_error_code(shell, err_code)) {
-		print_set_error(shell, "soft-start config");
-		return 0;
-	}
-
-	switch (config_type) {
-	case LDSW_SOFT_START_PARAM_ENABLE:
-		print_success(shell, result.bvalue, UNIT_TYPE_NONE);
-		break;
-	case LDSW_SOFT_START_PARAM_CURRENT:
-		print_success(shell, result.uvalue, UNIT_TYPE_MILLIAMPERE);
-		break;
-	}
-	return 0;
-}
-
-static int ldsw_soft_start_config_get(const struct shell *shell, size_t argc, char **argv,
-				      ldsw_soft_start_config_param_t config_type)
-{
-	args_info_t args_info = { .expected_args = 1,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	npmx_ldsw_soft_start_config_t soft_start_config;
-	npmx_error_t err_code = npmx_ldsw_soft_start_config_get(ldsw_instance, &soft_start_config);
-	if (!check_error_code(shell, err_code)) {
-		print_get_error(shell, "soft-start config");
-		return 0;
-	}
-
-	switch (config_type) {
-	case LDSW_SOFT_START_PARAM_ENABLE:
-		print_value(shell, soft_start_config.enable, UNIT_TYPE_NONE);
-		break;
-	case LDSW_SOFT_START_PARAM_CURRENT:
-		uint32_t soft_start_current_ma;
-		if (!npmx_ldsw_soft_start_current_convert_to_ma(soft_start_config.current,
-								&soft_start_current_ma)) {
-			print_convert_error(shell, "soft-start current", "milliamperes");
-			return 0;
-		}
-		print_value(shell, soft_start_current_ma, UNIT_TYPE_MILLIAMPERE);
-		break;
-	}
-	return 0;
-}
-
-static int cmd_ldsw_soft_start_current_set(const struct shell *shell, size_t argc, char **argv)
-{
-	return ldsw_soft_start_config_set(shell, argc, argv, LDSW_SOFT_START_PARAM_CURRENT);
-}
-
-static int cmd_ldsw_soft_start_current_get(const struct shell *shell, size_t argc, char **argv)
-{
-	return ldsw_soft_start_config_get(shell, argc, argv, LDSW_SOFT_START_PARAM_CURRENT);
-}
-
-static int cmd_ldsw_soft_start_enable_set(const struct shell *shell, size_t argc, char **argv)
-{
-	return ldsw_soft_start_config_set(shell, argc, argv, LDSW_SOFT_START_PARAM_ENABLE);
-}
-
-static int cmd_ldsw_soft_start_enable_get(const struct shell *shell, size_t argc, char **argv)
-{
-	return ldsw_soft_start_config_get(shell, argc, argv, LDSW_SOFT_START_PARAM_ENABLE);
-}
-
-static int cmd_ldsw_status_set(const struct shell *shell, size_t argc, char **argv)
-{
-	args_info_t args_info = { .expected_args = 2,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-					  [1] = { SHELL_ARG_TYPE_BOOL_VALUE, "status" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	bool ldsw_status = args_info.arg[1].result.bvalue;
-	npmx_ldsw_task_t ldsw_task = ldsw_status ? NPMX_LDSW_TASK_ENABLE : NPMX_LDSW_TASK_DISABLE;
-	npmx_error_t err_code = npmx_ldsw_task_trigger(ldsw_instance, ldsw_task);
-	if (!check_error_code(shell, err_code)) {
-		print_set_error(shell, "LDSW status");
-		return 0;
-	}
-
-	print_success(shell, ldsw_status, UNIT_TYPE_NONE);
-	return 0;
-}
-
-static int cmd_ldsw_status_get(const struct shell *shell, size_t argc, char **argv)
-{
-	args_info_t args_info = { .expected_args = 1,
-				  .arg = {
-					  [0] = { SHELL_ARG_TYPE_UINT32_INDEX, "LDSW" },
-				  } };
-	if (!arguments_check(shell, argc, argv, &args_info)) {
-		return 0;
-	}
-
-	npmx_ldsw_t *ldsw_instance = ldsw_instance_get(shell, args_info.arg[0].result.uvalue);
-	if (ldsw_instance == NULL) {
-		return 0;
-	}
-
-	uint8_t ldsw_idx = (uint8_t)args_info.arg[0].result.uvalue;
-	uint8_t status_mask;
-	npmx_error_t err_code = npmx_ldsw_status_get(ldsw_instance, &status_mask);
-	uint8_t check_mask = ldsw_idx == 0 ? (NPMX_LDSW_STATUS_POWERUP_LDSW_1_MASK |
-					      NPMX_LDSW_STATUS_POWERUP_LDO_1_MASK) :
-					     (NPMX_LDSW_STATUS_POWERUP_LDSW_2_MASK |
-					      NPMX_LDSW_STATUS_POWERUP_LDO_2_MASK);
-
-	if (!check_error_code(shell, err_code)) {
-		print_get_error(shell, "LDSW status");
-		return 0;
-	}
-
-	print_value(shell, !!(status_mask & check_mask), UNIT_TYPE_NONE);
-	return 0;
 }
 
 static int cmd_led_mode_set(const struct shell *shell, size_t argc, char **argv)
@@ -1418,71 +925,6 @@ static int cmd_vbusin_status_connected_get(const struct shell *shell, size_t arg
 	return 0;
 }
 
-/* Creating subcommands (level 3 command) array for command "ldsw active_discharge". */
-SHELL_STATIC_SUBCMD_SET_CREATE(sub_ldsw_active_discharge,
-			       SHELL_CMD(set, NULL, "Set active discharge status",
-					 cmd_ldsw_active_discharge_set),
-			       SHELL_CMD(get, NULL, "Get active discharge status",
-					 cmd_ldsw_active_discharge_get),
-			       SHELL_SUBCMD_SET_END);
-
-/* Creating dictionary subcommands (level 3 command) array for command "ldsw gpio". */
-SHELL_STATIC_SUBCMD_SET_CREATE(sub_ldsw_gpio,
-			       SHELL_CMD(set, NULL, "Set LDSW GPIO", cmd_ldsw_gpio_set),
-			       SHELL_CMD(get, NULL, "Get LDSW GPIO", cmd_ldsw_gpio_get),
-			       SHELL_SUBCMD_SET_END);
-
-/* Creating dictionary subcommands (level 3 command) array for command "ldsw ldo_voltage". */
-SHELL_STATIC_SUBCMD_SET_CREATE(sub_ldsw_ldo_voltage,
-			       SHELL_CMD(set, NULL, "Set LDO voltage", cmd_ldsw_ldo_voltage_set),
-			       SHELL_CMD(get, NULL, "Get LDO voltage", cmd_ldsw_ldo_voltage_get),
-			       SHELL_SUBCMD_SET_END);
-
-/* Creating dictionary subcommands (level 3 command) array for command "ldsw mode". */
-SHELL_STATIC_SUBCMD_SET_CREATE(sub_ldsw_mode,
-			       SHELL_CMD(set, NULL, "Set LDSW mode", cmd_ldsw_mode_set),
-			       SHELL_CMD(get, NULL, "Get LDSW mode", cmd_ldsw_mode_get),
-			       SHELL_SUBCMD_SET_END);
-
-/* Creating subcommands (level 4 command) array for command "ldsw soft_start current". */
-SHELL_STATIC_SUBCMD_SET_CREATE(sub_ldsw_soft_start_current,
-			       SHELL_CMD(set, NULL, "Set soft-start current",
-					 cmd_ldsw_soft_start_current_set),
-			       SHELL_CMD(get, NULL, "Get soft-start current",
-					 cmd_ldsw_soft_start_current_get),
-			       SHELL_SUBCMD_SET_END);
-
-/* Creating subcommands (level 4 command) array for command "ldsw soft_start enable". */
-SHELL_STATIC_SUBCMD_SET_CREATE(sub_ldsw_soft_start_enable,
-			       SHELL_CMD(set, NULL, "Set soft-start enable",
-					 cmd_ldsw_soft_start_enable_set),
-			       SHELL_CMD(get, NULL, "Get soft-start enable",
-					 cmd_ldsw_soft_start_enable_get),
-			       SHELL_SUBCMD_SET_END);
-
-/* Creating dictionary subcommands (level 3 command) array for command "ldsw soft_start". */
-SHELL_STATIC_SUBCMD_SET_CREATE(sub_ldsw_soft_start,
-			       SHELL_CMD(current, &sub_ldsw_soft_start_current,
-					 "Soft-start current", NULL),
-			       SHELL_CMD(enable, &sub_ldsw_soft_start_enable, "Soft-start enable",
-					 NULL),
-			       SHELL_SUBCMD_SET_END);
-
-/* Creating dictionary subcommands (level 3 command) array for command "ldsw status". */
-SHELL_STATIC_SUBCMD_SET_CREATE(sub_ldsw_status,
-			       SHELL_CMD(set, NULL, "Set LDSW status", cmd_ldsw_status_set),
-			       SHELL_CMD(get, NULL, "Get LDSW status", cmd_ldsw_status_get),
-			       SHELL_SUBCMD_SET_END);
-
-/* Creating subcommands (level 2 command) array for command "ldsw". */
-SHELL_STATIC_SUBCMD_SET_CREATE(
-	sub_ldsw, SHELL_CMD(active_discharge, &sub_ldsw_active_discharge, "Active discharge", NULL),
-	SHELL_CMD(gpio, &sub_ldsw_gpio, "Select GPIO used as LDSW's on/off", NULL),
-	SHELL_CMD(ldo_voltage, &sub_ldsw_ldo_voltage, "LDO voltage", NULL),
-	SHELL_CMD(mode, &sub_ldsw_mode, "LDSW mode", NULL),
-	SHELL_CMD(soft_start, &sub_ldsw_soft_start, "LDSW soft-start", NULL),
-	SHELL_CMD(status, &sub_ldsw_status, "LDSW status", NULL), SHELL_SUBCMD_SET_END);
-
 /* Creating subcommands (level 3 command) array for command "led mode". */
 SHELL_STATIC_SUBCMD_SET_CREATE(sub_led_mode, SHELL_CMD(set, NULL, "Set LED mode", cmd_led_mode_set),
 			       SHELL_CMD(get, NULL, "Get LED mode", cmd_led_mode_get),
@@ -1656,7 +1098,6 @@ SHELL_STATIC_SUBCMD_SET_CREATE(
 	sub_vbusin, SHELL_CMD(current_limit, &sub_vbusin_current_limit, "Current limit", NULL),
 	SHELL_CMD(status, &sub_vbusin_status, "Status", NULL), SHELL_SUBCMD_SET_END);
 
-SHELL_SUBCMD_ADD((npmx), ldsw, &sub_ldsw, "LDSW", NULL, 1, 0);
 SHELL_SUBCMD_ADD((npmx), led, &sub_led, "LED", NULL, 1, 0);
 SHELL_SUBCMD_ADD((npmx), pof, &sub_pof, "POF", NULL, 1, 0);
 SHELL_SUBCMD_ADD((npmx), reset, NULL, "Reset device", cmd_reset, 1, 0);
